requirements.txt
pandas>=2.0.0
azure-storage-blob>=12.14.0
requests>=2.28.0
xml.etree.ElementTree
datetime
collections
os
io
re 